### 2020-05-22 #2208 @whedon
[PRE REVIEW]: StoSpa2: A C++ software package for stochastic simulations of spatially extended systems
**Submitting author:** @BartoszBartmanski (<a href="http://orcid.org/0000-0002-0495-9150">Bartosz Bartmanski</a>)
**Repository:** <a href="https://github.com/BartoszBartmanski/StoSpa2" target ="_blank">https://github.com/BartoszBartmanski/StoSpa2</a>
**Version:** v2.0.27
