### 2020-06-04 #2208 @pdebuyl
@mbkumar could you review the JOSS article StoSpa2 ?
