### 2020-06-04 #2208 @pdebuyl
@arosen93 could you review the JOSS article StoSpa2 ?
