### 2020-06-04 #2208 @pdebuyl
@CFGrote could you review the JOSS article StoSpa2 ?
