### 2020-06-05 #2293 @whedon
[REVIEW]: StoSpa2: A C++ software package for stochastic simulations of spatially extended systems
**Submitting author:** @BartoszBartmanski (<a href="http://orcid.org/0000-0002-0495-9150">Bartosz Bartmanski</a>)
**Repository:** <a href="https://github.com/BartoszBartmanski/StoSpa2" target ="_blank">https://github.com/BartoszBartmanski/StoSpa2</a>
**Version:** 2.0.29

@CFGrote and @mbkumar , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
