### 2020-06-05 #2293 @pdebuyl
@CFGrote , @mbkumar , make sure to accept the invitation to the reviewers group and to have a look at the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review process will happen in this issue page, so questions to the author or to me can be added as comments here.
