### 2020-06-06 BartoszBartmanski/StoSpa2#2 @mbkumar
cme_example.svg file not available
The cme_example.svg file mentioned in the README.md is not referenced properly.

#openjournals/joss-reviews#2293
