### 2020-06-06 BartoszBartmanski/StoSpa2#3 @mbkumar
pip install not working
Not able to install pystospa with pip install.

The following error was thrown

```bash
> pip install pystospa
Collecting pystospa
  Downloading pystospa-2.0.28.tar.gz (752 kB)
     |████████████████████████████████| 752 kB 2.3 MB/s
     command: /home/mbkumar/anaconda3/envs/stospa2/bin/python -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-5f1aitqd/pystospa/setup.py'"'"'; __file__='"'"'/tmp/pip-install-5f1aitqd/pystospa/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /tmp/pip-install-5f1aitqd/pystospa/pip-egg-info
         cwd: /tmp/pip-install-5f1aitqd/pystospa/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-install-5f1aitqd/pystospa/setup.py", line 4, in <module>
        from skbuild import setup
    ModuleNotFoundError: No module named 'skbuild'
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
```

I think scikit-build has to be specified has a dependency.

#openjournals/joss-reviews#2293
