### 2020-06-06 BartoszBartmanski/StoSpa2#4 @mbkumar
circular dependency on scikit-build in the setup process
StoSpa2 can not be installed with the way the setup.py file is written unless someone installs scikit-build before trying to install pystospa with either `pip install pystospa` or `python setup.py install`.

Either update the documentation or remove the cyclic dependency on scikit-build.

#openjournals/joss-reviews#2293
