+++ setup.py
@@ -28,6 +28,6 @@ setup(
     platforms=["Linux"],
     cmake_args=cmake_args,
     packages=['pystospa'],
-    setup_requires=["setuptools", "wheel", "scikit-build >= 0.10.0", "cmake >= 3.5"],
+    setup_requires=["setuptools", "wheel", "cmake >= 3.5"],
     zip_safe=False,
 )
