### 2020-06-06 c47750a
1) Updated installation instructions in the README.md and ReadTheDocs documentation. 2) Corrected file path in the README.md.
