### 2020-06-06 BartoszBartmanski/StoSpa2#2 @BartoszBartmanski
Thank you for spotting that. The path to cme_example.svg in README.md has been corrected
