### 2020-06-06 BartoszBartmanski/StoSpa2#3 @BartoszBartmanski
I added instructions to install scikit-build before installing pystospa to README on github and the documentation on readthedocs
