### 2020-06-06 BartoszBartmanski/StoSpa2#4 @BartoszBartmanski
I removed the circular dependency on scikit-build by removing it from install_requires argument within setup.py and adding an instruction to README and the online documentation to install scikit-build before installing pystospa
