### 2020-06-06 BartoszBartmanski/StoSpa2#5 @mbkumar
Optional dependency on matplotlib
Though not a strict dependency, matplotlib is needed to run the python example.  I'd suggest to mention the dependency in the README or in fact add it to setup.py in order not to trip beginners. Based on my experience, you will encounter many users not at all familiar with python.

#openjournals/joss-reviews#2293
