### 2020-06-06 BartoszBartmanski/StoSpa2#6 @mbkumar
State of the field in the manuscript
I don't see any section on the status of the field and whether any competing software packages exist or not.

#openjournals/joss-reviews#2293
