### 2020-06-06 BartoszBartmanski/StoSpa2#7 @mbkumar
Unittests and feedback in docs
1) The docs or the README.md lack a section on how to run unit tests to make sure the code is correctly compiled and running. While experienced programmers can pick it up from the code layout, beginners can trip.
2) At present unit tests can be evaluated only if one compiles the C++ version.  Can the package support python based test packages such as nose to make it convenient for python users?
3) Add a few lines in the documentation on how to file bug reports

#openjournals/joss-reviews#2293
