### 2020-06-08 2fb4df2
1) Added information about dependencies for running examples to README file and the documenation 2) Added feedback section to the documentation 3) Added unit tests for pystospa
