+++ docs/src/feedback.rst
@@ -1,0 +1,11 @@
+
+Feedback
+########
+
+Report bugs
+===========
+
+Please submit any bugs as an issue to https://github.com/BartoszBartmanski/StoSpa2/issues.
+
+It would be greatly appreciated if the operating system used and/or the environment
+used are specified when reporting bugs.
