+++ .travis.yml
@@ -195,3 +195,4 @@ script:
     - ./tests/unittests
     - cd ../
     - python setup.py install
+    - python tests/test_pystospa.py
