### 2020-06-08 BartoszBartmanski/StoSpa2#7 @BartoszBartmanski
Thanks for suggesting this. A script to do unit-testing on pystospa has been added to tests directory. There are now two additional sections within the documentation: one about unit-testing and one about bug reports.
