### 2020-06-08 BartoszBartmanski/StoSpa2#5 @BartoszBartmanski
A note has been added in the README and the documentation that numpy and matplotlib are needed to run the python examples, as well as a command to install those dependencies if one wishes to run those examples. I chose not to add these dependencies to setup.py to keep the number of dependencies as low as possible.
