### 2020-06-10 a9c077f
Added a section to the JOSS paper.
