+++ docs/joss_paper/paper.md
@@ -36,6 +36,10 @@ To make the software user-friendly, the application programming interface (API) 
 
 The continuous integration platform TravisCI (https://travis-ci.org/) is used to make sure that any changes in the code-base do not break any functionality of `StoSpa2`. Both the installation and the functionality are tested on Linux and OSX operating systems.
 
+# Status of the field
+
+The CME and RDME frameworks are used to model various phenomena in fields such as chemistry, epidemiology and systems biology. The goal behind making `StoSpa2` is to facilitate easy and fast simulations of systems modelled using the CME and RDME. There are some alternative software packages that can be used, for example, URDME [@drawert2012urdme] and MesoRD [@fange2012mesord], however these both have dependencies that can be a barrier to installation and use; URDME has proprietary software as dependencies, while MesoRD has a large number of dependencies that make it challenging to install. StochSS [@drawert2016stochastic] is another software package, and provides a docker container and a web interface. However, docker containers require special privileges to run, which not every user may have, and the web interface does not allow high throughput execution of simulations. Hence, we developed `StoSpa2`, which relies upon as few dependencies as possible and is designed to be easily installed. Furthermore, our software can simulate dynamics on a uniformly growing domain using the Extrande method [@Voliotis2016], while the alternatives cannot.
+
 # Installation
 
 The code for `StoSpa2` is hosted on GitHub (https://github.com/BartoszBartmanski/StoSpa2) with installation instructions contained in the `README.md` file. The first way of installing `StoSpa2` involves downloading the source code from the GitHub repository and compiling the C++ code according to the instructions in the `README.md` file. However, an easier alternative is to use the Python package manager, pip, to download the Python binding of `StoSpa2`, `pystospa`, from https://pypi.org/project/pystospa/ and install it appropriately. All details of the installation, as well as more information, are contained in the documentation website (https://stospa2.readthedocs.io/en/latest/).
