### 2020-06-10 BartoszBartmanski/StoSpa2#6 @BartoszBartmanski
Thank you for this comment. I've updated paper.md with an additional section on competing software.
