### 2020-06-15 BartoszBartmanski/StoSpa2#8 @CFGrote
Cannot run the example as per docs
I followed the C++ installation instructions:

```
$> ./setup_stospa.sh                                                  [master]
Already on 'master'
Your branch is up to date with 'origin/master'.
Already up to date.
-- The CXX compiler identification is GNU 8.3.0
-- Check for working CXX compiler: /usr/bin/c++
-- Check for working CXX compiler: /usr/bin/c++ -- works
-- Detecting CXX compiler ABI info
-- Detecting CXX compile features
-- Project version: 2.0.29
-- CMAKE_BUILD_TYPE: Release
-- CMAKE_CXX_FLAGS: -O3 -DNDEBUG
-- Found PythonInterp: /opt/anaconda3/2018.12/bin/python3.7 (found version "3.7.4")
-- Found PythonLibs: /opt/anaconda3/2018.12/lib/libpython3.7m.so
-- pybind11 v2.4.dev4
-- Performing Test HAS_FLTO
-- Performing Test HAS_FLTO - Success
-- LTO enabled
-- Could NOT find Doxygen (missing: DOXYGEN_EXECUTABLE)
-- Found Sphinx: /opt/anaconda3/2018.12/bin/sphinx-build
-- Could NOT find CMakeLists.txt within examples directory
-- Configuring done
-- Generating done
-- Build files have been written to: /home/grotec/Projects/StoSpa2/build
Scanning dependencies of target benchmark_diffusion
[ 10%] Building CXX object benchmarks/CMakeFiles/benchmark_diffusion.dir/benchmark_diffusion.cpp.o
[ 20%] Linking CXX executable benchmark_diffusion
[ 20%] Built target benchmark_diffusion
Scanning dependencies of target benchmark_schnakenberg
[ 30%] Building CXX object benchmarks/CMakeFiles/benchmark_schnakenberg.dir/benchmark_schnakenberg.cpp.o
[ 40%] Linking CXX executable benchmark_schnakenberg
[ 40%] Built target benchmark_schnakenberg
Scanning dependencies of target benchmark_cme
[ 50%] Building CXX object benchmarks/CMakeFiles/benchmark_cme.dir/benchmark_cme.cpp.o
[ 60%] Linking CXX executable benchmark_cme
[ 60%] Built target benchmark_cme
Scanning dependencies of target unittests
[ 70%] Building CXX object tests/CMakeFiles/unittests.dir/unittests.cpp.o
[ 80%] Linking CXX executable unittests
[ 80%] Built target unittests
Scanning dependencies of target pystospa
[ 90%] Building CXX object pystospa/CMakeFiles/pystospa.dir/src/pystospa.cpp.o
[100%] Linking CXX shared module pystospa.cpython-37m-x86_64-linux-gnu.so
[100%] Built target pystospa
```

Note the message about missing CMakeLists.txt in examples. The file exists, though.

Next, I tried to run the example:

```
$> cd build
$> ./example
 no such file or directory: ./example
```

#openjournals/joss-reviews#2293
