### 2020-06-15 #2293 @CFGrote
Can't run the example. See https://github.com/BartoszBartmanski/StoSpa2/issues/8 .
