### 2020-06-15 #2293 @CFGrote
The target audience is not clearly mentioned (neither in README nor in the documentation).
