### 2020-06-15 #2293 @CFGrote
I suggest to add short sections on how to contribute and how to seek support (i.e. link to gh issues) to the README.
