### 2020-06-15 #2293 @CFGrote
Some references in the manuscript are not correctly rendered in the autogenerated pdf but this may be a conversion problem. Please check the page number in Osborne et al 2017) and Isaacson 2009.
