### 2020-06-15 #2293 @CFGrote
Installation instructions:
Only after reading the manuscript, I understood that it is not required to compile the c++ code if one is only using the python interface. This should maybe be stated more clearly in the online installation instructions in README and docs.
