### 2020-06-15 63e7cac
1) Update documentation with outdated example in the documentation 2) Fixed a typo in CMakeLists.txt 3) Updated the Dependencies for clarity.
