### 2020-06-15 adcf5aa
1) Added a section on contributing via pull requests 2) Updated references for JOSS submission 3) Added clarification to installtion instuctions in Docs and README 4) Specified the target audience in the paper.md
