+++ docs/joss_paper/paper.bib
@@ -1,7 +1,7 @@
 
 @article{drawert2016stochastic,
   title={Stochastic simulation service: bridging the gap between the computational expert and the biologist},
-  author={Drawert, Brian and Hellander, Andreas and Bales, Ben and Banerjee, Debjani and Bellesia, Giovanni and Daigle Jr, Bernie J and Douglas, Geoffrey and Gu, Mengyuan and Gupta, Anand and Hellander, Stefan and others},
+  author={Drawert, B. and Hellander, A. and Bales, B. and Banerjee, D. and Bellesia, G. and Daigle Jr, B. J. and Douglas, G. and Gu, M. and Gupta, A, and Hellander, S. and others},
   journal={PLoS computational biology},
   volume={12},
   number={12},
@@ -11,7 +11,7 @@
 
 @article{drawert2012urdme,
   title={URDME: a modular framework for stochastic simulation of reaction-transport processes in complex geometries},
-  author={Drawert, Brian and Engblom, Stefan and Hellander, Andreas},
+  author={Drawert, B. and Engblom, S. and Hellander, A.},
   journal={BMC systems biology},
   volume={6},
   number={1},
@@ -22,7 +22,7 @@
 
 @article{fange2012mesord,
   title={MesoRD 1.0: Stochastic reaction-diffusion simulations in the microscopic limit},
-  author={Fange, David and Mahmutovic, Anel and Elf, Johan},
+  author={Fange, D. and Mahmutovic, A. and Elf, J.},
   journal={Bioinformatics},
   volume={28},
   number={23},
@@ -63,19 +63,14 @@
 }
 
 @article{Osborne2017,
-	Abstract = {The coordinated behaviour of populations of cells plays a central role in tissue growth and renewal. Cells react to their microenvironment by modulating processes such as movement, growth and proliferation, and signalling. Alongside experimental studies, computational models offer a useful means by which to investigate these processes. To this end a variety of cell-based modelling approaches have been developed, ranging from lattice-based cellular automata to lattice-free models that treat cells as point-like particles or extended shapes. However, it remains unclear how these approaches compare when applied to the same biological problem, and what differences in behaviour are due to different model assumptions and abstractions. Here, we exploit the availability of an implementation of five popular cell-based modelling approaches within a consistent computational framework, Chaste (http://www.cs.ox.ac.uk/chaste). This framework allows one to easily change constitutive assumptions within these models. In each case we provide full details of all technical aspects of our model implementations. We compare model implementations using four case studies, chosen to reflect the key cellular processes of proliferation, adhesion, and short- and long-range signalling. These case studies demonstrate the applicability of each model and provide a guide for model usage.},
-	Author = {Osborne, J. M. and Fletcher, A. G. and Pitt-Francis, J. M. and Maini, P. K. and Gavaghan, D. J.},
-	Doi = {10.1371/journal.pcbi.1005387},
-	Isbn = {1111111111},
-	Issn = {15537358},
-	Journal = {PLoS Computational Biology},
-	Keywords = {off{\_}lattice{\_}approaches},
-	Number = {2},
-	Pages = {1--34},
-	Pmid = {28192427},
-	Title = {{Comparing individual-based approaches to modelling the self-organization of multicellular tissues}},
-	Volume = {13},
-	Year = {2017},
+  title={Comparing individual-based approaches to modelling the self-organization of multicellular tissues},
+  author={Osborne, J. M. and Fletcher, A. G. and Pitt-Francis, J. M. and Maini, P. K. and Gavaghan, D. J.},
[…]
