+++ docs/src/install.rst
@@ -2,8 +2,7 @@
 Installation
 ############
 
-There are two ways of using StoSpa2: with Python or with C++.
-
+There are two ways of using StoSpa2: with Python or with C++. To install `pystospa`, we only need to follow the installation instructions found in the Python section below and likewise, to compile the C++ code, we only need to follow the instructions in the C++ section.
 
 Prerequisites
 =============
@@ -38,6 +37,8 @@ installing pystospa, which can be done using the following command:
 .. code-block:: bash
 
     pip install scikit-build
+
+Furthermore, there is no need to compile the C++ code yourself - assuming the dependencies are met, pip and/or scikit-build compiles the C++ part of the software for us.
 
 Easy way
 --------
@@ -108,7 +109,7 @@ After the above command we can run the example executable (contained in your cur
 
         ./examples/cme_example
 
-which generates an `cme_example.dat` file.
+which generates an `cme_example.dat` file in your current working directory.
 
 This is the basic installation. There are more options, such as installation that uses the Boost library,
 covered in Branches section below.
