### 2020-06-15 BartoszBartmanski/StoSpa2#8 @BartoszBartmanski
Thanks for this comment. That part in the docs was a bit out of date. I just updated the docs and fixed a typo in the main CMakeLists.txt. The compiled examples should work now.
