### 2020-06-15 #2293 @BartoszBartmanski
@CFGrote thank you for these comments. I just pushed two commits to the repository for StoSpa2, which hopefully address all of the above mentioned points.
