### 2020-06-15 cd20b13
1) Added Feedback section to README, 2) Improved bibliography for JOSS paper
