+++ docs/joss_paper/paper.bib
@@ -1,131 +1,108 @@
 
 @article{drawert2016stochastic,
-  title={Stochastic simulation service: bridging the gap between the computational expert and the biologist},
-  author={Drawert, B. and Hellander, A. and Bales, B. and Banerjee, D. and Bellesia, G. and Daigle Jr, B. J. and Douglas, G. and Gu, M. and Gupta, A, and Hellander, S. and others},
-  journal={PLoS computational biology},
-  volume={12},
-  number={12},
-  year={2016},
-  publisher={Public Library of Science}
+    title={Stochastic simulation service: bridging the gap between the computational expert and the biologist},
+    author={Drawert, B. and Hellander, A. and Bales, B. and Banerjee, D. and Bellesia, G. and Daigle Jr, B. J. and Douglas, G. and Gu, M. and Gupta, A, and Hellander, S. and others},
+    journal={PLoS computational biology},
+    volume={12},
+    number={12},
+    year={2016},
+    publisher={Public Library of Science}
 }
 
 @article{drawert2012urdme,
-  title={URDME: a modular framework for stochastic simulation of reaction-transport processes in complex geometries},
-  author={Drawert, B. and Engblom, S. and Hellander, A.},
-  journal={BMC systems biology},
-  volume={6},
-  number={1},
-  pages={76},
-  year={2012},
-  publisher={BioMed Central}
+    title={URDME: a modular framework for stochastic simulation of reaction-transport processes in complex geometries},
+    author={Drawert, B. and Engblom, S. and Hellander, A.},
+    journal={BMC systems biology},
+    volume={6},
+    number={1},
+    pages={76},
+    year={2012},
+    publisher={BioMed Central}
 }
 
 @article{fange2012mesord,
-  title={MesoRD 1.0: Stochastic reaction-diffusion simulations in the microscopic limit},
-  author={Fange, D. and Mahmutovic, A. and Elf, J.},
-  journal={Bioinformatics},
-  volume={28},
-  number={23},
-  pages={3155--3157},
-  year={2012},
-  publisher={Oxford University Press}
+    title={MesoRD 1.0: Stochastic reaction-diffusion simulations in the microscopic limit},
+    author={Fange, D. and Mahmutovic, A. and Elf, J.},
+    journal={Bioinformatics},
+    volume={28},
+    number={23},
+    pages={3155--3157},
+    year={2012},
+    publisher={Oxford University Press}
 }
 
 @article{Voliotis2016,
-	abstract = {Simulation of biomolecular networks is now indispensable for studying biological systems, from small reaction networks to large ensembles of cells. Here we present a novel approach for stochastic simulation of networks embedded in the dynamic environment of the cell and its surroundings. We thus sample trajectories of the stochastic process described by the chemical master equation with time-varying propensities. A comparative analysis shows that existing approaches can either fail dramatically, or else can impose impractical computational burdens due to numerical integration of reaction propensities, especially when cell ensembles are studied. Here we introduce the Extrande method which, given a simulated time course of dynamic network inputs, provides a conditionally exact and several orders-of-magnitude faster simulation solution. The new approach makes it feasible to demonstrate, using decision-making by a large population of quorum sensing bacteria, that robustness to fluctuations from upstream signaling places strong constraints on the design of networks determining cell fate. Our approach has the potential to significantly advance both understanding of molecular systems biology and design of synthetic circuits.},
-	archivePrefix = {arXiv},
-	arxivId = {1511.01268},
-	author = {Voliotis, M. and Thomas, P. and Grima, R. and Bowsher, C. G.},
[…]
