+++ README.md
@@ -200,3 +200,21 @@ while for three voxels, with a single species of molecules the output will look 
 .
 .
 ```
+
+## Feedback
+
+### Report bugs
+
+Please submit any bugs as an issue to https://github.com/BartoszBartmanski/StoSpa2/issues.
+
+It would be greatly appreciated if the operating system used and/or the environment
+used are specified when reporting bugs.
+
+### Contributing
+
+If anyone wishes to contribute to this project, please submit a pull request to the
+`develop` branch in the repository for this software. We merge the `develop` with the
+`master` branch once a new feature on the `develop` branch has been tested.
+
+A guide on how to submit pull requests can be found on
+https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/creating-a-pull-request.
