### 2020-06-16 #2293 @pdebuyl
@BartoszBartmanski Some references still don't appear correctly, as you can only use one bib reference per square brackets. Can you fix that?
