### 2020-06-16 cfa634e
Separated the references in the JOSS paper with brackets to hopefully fix the rendering problem
