+++ docs/joss_paper/paper.bib
@@ -54,14 +54,18 @@
 
 
 @article{Osborne2017,
-    title={Comparing individual-based approaches to modelling the self-organization of multicellular tissues},
-    author={Osborne, J. M. and Fletcher, A. G. and Pitt-Francis, J. M. and Maini, P. K. and Gavaghan, D. J.},
-    journal={PLoS Computational Biology},
-    volume={13},
-    number={2},
-    pages={e1005387},
-    year={2017},
-    publisher={Public Library of Science}
+    author = {Osborne, J. M. and Fletcher, A. G. and Pitt-Francis, J. M. and Maini, P. K. and Gavaghan, D. J.},
+    journal = {PLOS Computational Biology},
+    publisher = {Public Library of Science},
+    title = {Comparing individual-based approaches to modelling the self-organization of multicellular tissues},
+    year = {2017},
+    month = {02},
+    volume = {13},
+    url = {https://doi.org/10.1371/journal.pcbi.1005387},
+    pages = {1-34},
+    abstract = {Author summary In combination with molecular and live-imaging techniques, computational modelling plays an increasingly important role in the study of tissue growth and renewal. To this end a variety of cell-based modelling approaches have been developed, ranging in complexity from lattice-based cellular automata to lattice-free models that treat cells as point-like particles or extended shapes. However, it remains unclear how these approaches compare when applied to the same biological problem, and under which circumstances each approach is valid. Here we implement five classes of such model in a consistent computational framework, Chaste. We apply each model to four simulation studies, chosen to illustrate how the cellular processes such as proliferation, adhesion, and short- and long-range signalling may be implemented in each model. These case studies demonstrate the applicability of each model and highlight where one may expect to see qualitative differences between model behaviours. Taken together, these findings provide a guide for model usage.},
+    number = {2},
+    doi = {10.1371/journal.pcbi.1005387}
 }
 
 @article{VanLiedekerke2015,
