### 2020-06-16 #2293 @BartoszBartmanski
Sorry, I did not realise that it was one reference per square brackets. I just updated the repository with those changes. Thank you for clarifying that.
