### 2020-06-16 #2293 @danielskatz
Hi
1 - you can have multiple references inside one set of square brackets, but they need to separated with ;s, not ,s - they may also not support a space after the ; but I am not sure.  Try [@a;@b] for example.
2 - commands to whedon need to be the very start of a comment
