### 2020-06-16 #2293 @pdebuyl
Thank you @danielskatz

Sorry @BartoszBartmanski , it is indeed better to have the grouped citations as originally intended.
