### 2020-06-16 07cef8d
Updated citations in JOSS paper
