### 2020-06-16 #2293 @BartoszBartmanski
Thank you @danielskatz for both suggestions

@danielskatz I just updated the citations in the repository. Let's see whether it's fixed
