### 2020-06-17 08fd4ec
1) Added primary audience to README and docs. 2) Fixed a typo in README
