### 2020-06-17 #2293 @BartoszBartmanski
@CFGrote Thank you for spotting Examples section misalignment. A typo in the README caused this section title to be interpreted as a code block. I also added target audience to both the README and the docs in the most recent commit.
