### 2020-06-17 #2293 @CFGrote
@BartoszBartmanski, very good. I'm happy with the repo now and recommend publication as is.
