### 2020-06-17 #2293 @CFGrote
@pdebuyl I'm done with the review, my recommendation is: publish as is. Thanks for the experience, this was my first JOSS review.
