### 2020-06-17 #2293 @pdebuyl
@BartoszBartmanski can you add the doi for all references in the bib file?
