### 2020-06-17 ade719b
Added doi's to references for the JOSS paper.
