+++ docs/joss_paper/paper.bib
@@ -17,7 +17,8 @@
     number={1},
     pages={76},
     year={2012},
-    publisher={BioMed Central}
+    publisher={BioMed Central},
+    doi={10.1186/1752-0509-6-76}
 }
 
 @article{fange2012mesord,
@@ -28,7 +29,8 @@
     number={23},
     pages={3155--3157},
     year={2012},
-    publisher={Oxford University Press}
+    publisher={Oxford University Press},
+    doi={10.1093/bioinformatics/bts584}
 }
 
 @article{Voliotis2016,
@@ -38,7 +40,8 @@
     volume={12},
     number={6},
     year={2016},
-    publisher={Public Library of Science}
+    publisher={Public Library of Science},
+    doi={10.1371/journal.pcbi.1004923}
 }
 
 @article{Hellander2017,
@@ -49,7 +52,8 @@
     number={6},
     pages={064101},
     year={2017},
-    publisher={AIP Publishing LLC}
+    publisher={AIP Publishing LLC},
+    doi={10.1063/1.4975167}
 }
 
 
@@ -76,7 +80,8 @@
     number={4},
     pages={401--444},
     year={2015},
-    publisher={Springer}
+    publisher={Springer},
+    doi={10.1007/s40571-015-0082-3}
 }
 
 
@@ -88,7 +93,8 @@
     number={1},
     pages={77--111},
     year={2009},
-    publisher={SIAM}
+    publisher={SIAM},
+    doi={10.1137/070705039}
 }
 
 @article{Isaacson2013,
@@ -99,7 +105,8 @@
     number={5},
     pages={054101},
     year={2013},
-    publisher={American Institute of Physics}
+    publisher={American Institute of Physics},
+    doi={10.1063/1.4816377}
 }
 
 
