### 2020-06-17 #2293 @BartoszBartmanski
@pdebuyl I've just updated the reference file with doi
