### 2020-06-17 #2293 @pdebuyl
@BartoszBartmanski thank you for the merges.

Could you make a tagged release and archive it (on Zenodo). When you've done that, report the version number and archive DOI here please. Please make sure the archive deposit has the correct metadata (title and author list) to match the paper.

If you could also confirm the version number, it's currently v2.0.27 but was updated to 2.0.29
