### 2020-06-17 #2293 @BartoszBartmanski
@pdebuyl The version number is 2.0.29 and DOI is 10.5281/zenodo.3899466
