### 2020-06-18 #2293 @BartoszBartmanski
Thank you @pdebuyl @CFGrote @mbkumar for your time and your comments.

This was my first submission to JOSS and it was a very pleasant experience submitting a paper.
