### 2020-06-18 #2293 @BartoszBartmanski
@pdebuyl Can I ask who has to run the command "whedon accept deposit=true" for the pull request to be merged with joss (https://github.com/openjournals/joss-papers/pull/1495)?
