### 2020-06-18 #2293 @pdebuyl
@BartoszBartmanski it is also my experience that JOSS reviews are nicer than typical ones :-) Thank you for the feedback.

You don't have to perform any task, nor do I. The "deposit=true"  action is for someone from @openjournals/joss-eics
