### 2020-06-19 #2293 @BartoszBartmanski
@pdebuyl Is it possible to change the zenodo doi for this submission?

There was a mix up with some people at zenodo and the archive has been deleted. I can set up a new archive and give the new doi. Would that be ok?

Sorry for the trouble.
