### 2020-06-19 #2293 @danielskatz
👋 @pdebuyl  - this paper is too long, which I'm going to let go at this point, but for future papers you edit, a lot of the examples should be in the documentation, with the paper just pointing to the documentation, not including this much detail.  See https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
