### 2020-06-20 #2293 @pdebuyl
@danielskatz thank you for the finalization. I will keep the length in mind for my future editorial assignments.
